# anp_open_sdk/anp_users/openapi_did%3Awba%3Alocalhost%253A9527%3Awba%3Auser%3A5fea49e183c6c211.yaml
openapi: 3.0.0
info:
  title: Agent did:wba:localhost%3A9527:wba:user:5fea49e183c6c211 API Documentation
  version: 1.0.0
paths:
  /wba/auth:
    get:
      summary: DID WBA authentication endpoint
      operationId: get__wba_auth
      parameters:
      - name: req_did
        in: query
        required: false
        schema:
          type: string
        default: demo_caller
      responses:
        '200':
          description: 成功响应
          content:
            application/json:
              schema:
                type: object
  /wba/user/{user_id}/did.json:
    get:
      summary: Get DID document
      operationId: get__wba_user_user_id_did.json
      parameters:
      - name: req_did
        in: query
        required: false
        schema:
          type: string
        default: demo_caller
      responses:
        '200':
          description: 成功响应
          content:
            application/json:
              schema:
                type: object
  /wba/user/{user_id}/ad.json:
    get:
      summary: Get agent description
      operationId: get__wba_user_user_id_ad.json
      parameters:
      - name: req_did
        in: query
        required: false
        schema:
          type: string
        default: demo_caller
      responses:
        '200':
          description: 成功响应
          content:
            application/json:
              schema:
                type: object
  /wba/user/{resp_did}/{yaml_file_name}.yaml:
    get:
      summary: Get agent OpenAPI YAML
      operationId: get__wba_user_resp_did_yaml_file_name.yaml
      parameters:
      - name: req_did
        in: query
        required: false
        schema:
          type: string
        default: demo_caller
      responses:
        '200':
          description: 成功响应
          content:
            application/json:
              schema:
                type: object
  /wba/user/{resp_did}/{jsonrpc_file_name}.json:
    get:
      summary: Get agent JSON-RPC
      operationId: get__wba_user_resp_did_jsonrpc_file_name.json
      parameters:
      - name: req_did
        in: query
        required: false
        schema:
          type: string
        default: demo_caller
      responses:
        '200':
          description: 成功响应
          content:
            application/json:
              schema:
                type: object
  /wba/hostuser/{user_id}/did.json:
    get:
      summary: Get Hosted DID document
      operationId: get__wba_hostuser_user_id_did.json
      parameters:
      - name: req_did
        in: query
        required: false
        schema:
          type: string
        default: demo_caller
      responses:
        '200':
          description: 成功响应
          content:
            application/json:
              schema:
                type: object
  /wba/hosted-did/request:
    post:
      summary: submit_hosted_did_request
      operationId: post__wba_hosted-did_request
      parameters:
      - name: req_did
        in: query
        required: false
        schema:
          type: string
        default: demo_caller
      responses:
        '200':
          description: 成功响应
          content:
            application/json:
              schema:
                type: object
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
  /wba/hosted-did/status/{request_id}:
    get:
      summary: check_hosted_did_status
      operationId: get__wba_hosted-did_status_request_id
      parameters:
      - name: req_did
        in: query
        required: false
        schema:
          type: string
        default: demo_caller
      responses:
        '200':
          description: 成功响应
          content:
            application/json:
              schema:
                type: object
  /wba/hosted-did/check/{requester_did_id}:
    get:
      summary: check_hosted_did_result
      operationId: get__wba_hosted-did_check_requester_did_id
      parameters:
      - name: req_did
        in: query
        required: false
        schema:
          type: string
        default: demo_caller
      responses:
        '200':
          description: 成功响应
          content:
            application/json:
              schema:
                type: object
  /wba/hosted-did/acknowledge/{result_id}:
    post:
      summary: acknowledge_hosted_did_result
      operationId: post__wba_hosted-did_acknowledge_result_id
      parameters:
      - name: req_did
        in: query
        required: false
        schema:
          type: string
        default: demo_caller
      responses:
        '200':
          description: 成功响应
          content:
            application/json:
              schema:
                type: object
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
  /wba/hosted-did/list:
    get:
      summary: list_hosted_dids
      operationId: get__wba_hosted-did_list
      parameters:
      - name: req_did
        in: query
        required: false
        schema:
          type: string
        default: demo_caller
      responses:
        '200':
          description: 成功响应
          content:
            application/json:
              schema:
                type: object
  /agent/message/{did}/post:
    post:
      summary: message_entry_post
      operationId: post__agent_message_did_post
      parameters:
      - name: req_did
        in: query
        required: false
        schema:
          type: string
        default: demo_caller
      responses:
        '200':
          description: 成功响应
          content:
            application/json:
              schema:
                type: object
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
  /agent/group/{did}/{group_id}/join:
    post:
      summary: join_group
      operationId: post__agent_group_did_group_id_join
      parameters:
      - name: group_id
        in: path
        required: true
        schema:
          type: string
      - name: req_did
        in: query
        required: false
        schema:
          type: string
        default: demo_caller
      responses:
        '200':
          description: 成功响应
          content:
            application/json:
              schema:
                type: object
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
  /agent/group/{did}/{group_id}/leave:
    post:
      summary: leave_group
      operationId: post__agent_group_did_group_id_leave
      parameters:
      - name: group_id
        in: path
        required: true
        schema:
          type: string
      - name: req_did
        in: query
        required: false
        schema:
          type: string
        default: demo_caller
      responses:
        '200':
          description: 成功响应
          content:
            application/json:
              schema:
                type: object
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
  /agent/group/{did}/{group_id}/message:
    post:
      summary: group_message
      operationId: post__agent_group_did_group_id_message
      parameters:
      - name: group_id
        in: path
        required: true
        schema:
          type: string
      - name: req_did
        in: query
        required: false
        schema:
          type: string
        default: demo_caller
      responses:
        '200':
          description: 成功响应
          content:
            application/json:
              schema:
                type: object
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
  /agent/group/{did}/{group_id}/connect:
    get:
      summary: group_connect
      operationId: get__agent_group_did_group_id_connect
      parameters:
      - name: group_id
        in: path
        required: true
        schema:
          type: string
      - name: req_did
        in: query
        required: false
        schema:
          type: string
        default: demo_caller
      responses:
        '200':
          description: 成功响应
          content:
            text/event-stream:
              schema:
                type: string
  /agent/group/{did}/{group_id}/members:
    post:
      summary: manage_group_members
      operationId: post__agent_group_did_group_id_members
      parameters:
      - name: group_id
        in: path
        required: true
        schema:
          type: string
      - name: req_did
        in: query
        required: false
        schema:
          type: string
        default: demo_caller
      responses:
        '200':
          description: 成功响应
          content:
            application/json:
              schema:
                type: object
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
    get:
      summary: get_group_members
      operationId: get__agent_group_did_group_id_members
      parameters:
      - name: group_id
        in: path
        required: true
        schema:
          type: string
      - name: req_did
        in: query
        required: false
        schema:
          type: string
        default: demo_caller
      responses:
        '200':
          description: 成功响应
          content:
            application/json:
              schema:
                type: object